Removed an useless marker check
This check also prevented wardens from being able to place markers without moving their mouse when right clicking. This was very confusing for wardens.

Moreover, it is impossible to get a negative value with GetVectorDistance (and it wouldn't be problematic as the radius of the circle is forced to a min&max).